refactor: split onboard prompt ui
